product model 修改

+++ Models/Product.cs
@@ -10,7 +10,7 @@ namespace BuildSchool.MvcSolution.OnlineStore.Models
     {
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public int UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public string Description { get; set; }
         public int CategoryID { get; set; }
         public string ProductImage { get; set; }
